se corrigio errores de eslint, se le coloco un confirm al trash listener y otros detalles pequeños

+++ src/views/post-display.js
@@ -1,10 +1,13 @@
+/* eslint-disable no-restricted-globals */
+/* eslint-disable no-alert */
 import { readData, deletePost } from "../firebase.js";
 
 export const newPost = (posts) => {
   const newPostContainer = document.querySelector("#postSection");
   newPostContainer.innerHTML = "";
   const feedContent = (e) => {
-    const asdf = Object.values(e);
+    const postsValues = Object.values(e);
+    // console.log(e);
     const feedHtml = `<div class="post-user">
             <div class="profile-container">
             <div class="pic-container">
@@ -13,15 +16,15 @@ export const newPost = (posts) => {
             <h3 class="title-profile">Lady Gaga</h3>
             </div>
             <div class="textarea-container">
-            <textarea class="post-theme" id="postTheme" cols="10" rows="1" readonly>${asdf[0].data.theme}</textarea>
-            <textarea class="post-message" id="postMessage" cols="15" rows="5" readonly>${asdf[0].data.message}</textarea>
+            <textarea class="post-theme" id="postTheme" cols="10" rows="1" readonly>${postsValues[0].data.theme}</textarea>
+            <textarea class="post-message" id="postMessage" cols="15" rows="5" readonly>${postsValues[0].data.message}</textarea>
             </div>
             <div class="icons-container">
             <div class="like-container">
             <img src="./resources/images/corazon.png" alt="like">
             </div>
               <div class="trash-container"> 
-              <button id="trash-icon" value=${asdf[0].id}><img src="./resources/images/trash.png" alt="trash"></button>
+              <button class="trash-icon" id="trashBtn" value=${postsValues[0].id}><img src="./resources/images/trash.png" alt="trash"></button>
               </div>
               <div class="edit-container">
               <img src="./resources/images/edit.png" alt="edit">
@@ -33,10 +36,14 @@ export const newPost = (posts) => {
   };
 
   posts.forEach(feedContent);
-  const deleteBtn = newPostContainer.querySelectorAll("#trash-icon");
+
+  const deleteBtn = newPostContainer.querySelectorAll("#trashBtn");
+
   deleteBtn.forEach((trash) => {
     trash.addEventListener("click", () => {
-      deletePost(trash.value);
+      if (confirm("¿estás segura que quieres eliminar el post?")) {
+        deletePost(trash.value);
+      }
     });
   });
   return newPostContainer;
